move BaOverlay to /components
